packages/ordeq-bigquery/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "setuptools_scm>=8.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ordeq_bigquery"
dynamic = ["version"]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    { name = "Niels Neerhoff" },
    { name = "Simon Brugman" },
    { name = "Nathan Bijleveld" },
]
maintainers = [
    { name = "Niels Neerhoff" },
    { name = "Simon Brugman" },
    { name = "Nathan Bijleveld" },
]
requires-python = ">=3.10"
dependencies = [
    "ordeq>=1.0.3",
    "google-cloud-bigquery==3.38.0",
    "pandas"
]
[tool.setuptools_scm]
root = "../../"
version_file = "src/ordeq_bigquery/_version.py"
tag_regex = "^ordeq-google-cloud/(?:[\\w-]+-)?(?P<version>[vV]?\\d+(?:\\.\\d+){0,2}[^\\+]*)(?:\\+.*)?$"
git_describe_command = "git describe --dirty --tags --long --match ordeq-google-cloud/*"

[tool.setuptools.packages.find]
namespaces = true
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
addopts = "--doctest-modules"

[tool.coverage.run]
omit = [
    "src/ordeq_bigquery/_version.py"
]
